+++ app/controllers/budgets_controller.rb
@@ -18,7 +18,7 @@ class BudgetsController < BaseController
     @activity = Activity.find(activity_id)
     authorize @activity
 
-    result = CreateBudget.new(activity: @activity).call(attributes: budget_params)
+    result = CreateBudget.new(activity: @activity).call(attributes: create_budget_params)
     @budget = result.object
 
     if result.success?
@@ -45,7 +45,7 @@ class BudgetsController < BaseController
 
     @activity = Activity.find(activity_id)
     result = UpdateBudget.new(budget: @budget, user: current_user)
-      .call(attributes: budget_params)
+      .call(attributes: update_budget_params)
 
     if result.success?
       flash[:notice] = t("action.budget.update.success")
@@ -67,6 +67,18 @@ class BudgetsController < BaseController
     redirect_to organisation_activity_path(@activity.organisation, @activity)
   end
 
+  def revisions
+    @activity = Activity.find(activity_id)
+    budget = Budget.find(params[:budget_id])
+
+    authorize budget
+
+    @audits = budget.audits
+
+    prepare_default_activity_trail(@activity)
+    add_breadcrumb t("breadcrumb.budget.revisions"), activity_budget_revisions_path(budget.parent_activity_id, budget)
+  end
+
   private
 
   def id
@@ -77,7 +89,7 @@ class BudgetsController < BaseController
     params[:activity_id]
   end
 
-  def budget_params
+  def create_budget_params
     params.require(:budget).permit(
       :budget_type,
       :value,
@@ -91,6 +103,7 @@ class BudgetsController < BaseController
     )
   end
 
-  def find_activity
+  def update_budget_params
+    params.require(:budget).permit(:value, :audit_comment)
   end
 end